Move auth from helper to package
