+++ tests/test_helpers.py
@@ -64,41 +64,3 @@ def test_db_url_from_hass_config():
         }
     }):
         assert helpers.db_url_from_hass_config('mock-path') == 'mock-url'
-
-
-def test_auth_from_hass_config():
-    """Test extracting users from hass config."""
-    mopen = mock_open(read_data=json.dumps({
-        'key': 'auth',
-        'version': 1,
-        'data': {
-            'users': [{
-                'id': 'mock-user-id',
-                'name': 'mock-user-name'
-            }],
-            'refresh_tokens': [{
-                'id': 'mock-refresh-id',
-                'user_id': 'mock-user-id',
-                'client_name': 'mock-client-name',
-                'client_id': 'mock-client-id',
-            }]
-        }
-    }))
-
-    with patch('detective.helpers.open', mopen):
-        auth = helpers.auth_from_hass_config('/some/path')
-
-    assert auth['users'] == {
-        'mock-user-id': {
-            'id': 'mock-user-id',
-            'name': 'mock-user-name'
-        }
-    }
-    assert auth['refresh_tokens_to_users'] == {
-        'mock-refresh-id': {
-            'id': 'mock-refresh-id',
-            'user': auth['users']['mock-user-id'],
-            'client_id': 'mock-client-id',
-            'client_name': 'mock-client-name',
-        }
-    }
